requirements.txt
discord.py
prometheus-client